+++ nel/src/3d/render_trav.cpp
@@ -1,7 +1,7 @@
 /** \file render_trav.cpp
  * TODO: File description
  *
- * $Id: render_trav.cpp,v 1.67 2006/05/31 12:03:14 boucher Exp $
+ * $Id: render_trav.cpp,v 1.67.6.1 2006/09/21 20:50:31 cado Exp $
  */
 
 /* Copyright, 2000 Nevrax Ltd.
@@ -1309,12 +1309,12 @@ void CRenderTrav::debugWaterModelMemory(const char *tag, bool dumpList)
 		{
 			// Before crash, do some log
 			nlwarning("******* WaterModelList crash after %s", tag);
-			nlwarning("Current: Ptr:%x. List:%x/%x", (uint32)dmp.Address, (uint32)dmp.ClippedPolyBegin, (uint32)dmp.ClippedPolyEnd);
+			nlwarning("Current: Ptr:%x. List:%x/%x", (ptrdiff_t)dmp.Address, (ptrdiff_t)dmp.ClippedPolyBegin, (ptrdiff_t)dmp.ClippedPolyEnd);
 			// Log also the list bkuped (to do comparisons)
 			for(uint i=0;i<_DebugWaterModelList.size();i++)
 			{
 				CWaterModelDump		&bkup= _DebugWaterModelList[i];
-				nlwarning("List%02d: Ptr:%x. Array:%x/%x", i, (uint32)bkup.Address, (uint32)bkup.ClippedPolyBegin, (uint32)bkup.ClippedPolyEnd);
+				nlwarning("List%02d: Ptr:%x. Array:%x/%x", i, (ptrdiff_t)bkup.Address, (ptrdiff_t)bkup.ClippedPolyBegin, (ptrdiff_t)bkup.ClippedPolyEnd);
 			}
 
 			// crash (assert not stop for clearness)
